Minor bug fixes
Consumer Protection BC updated their website so I’ve updated the “Confirm Film Rating” button. It’s no longer automated. This was hacky and unreliable to begin with and their new website has made it even harder, so instead I’ve added a preview box like the trailer and website links.

Fixed the trailer and website preview boxes. They were missing the open/close icons. I also updated the trailer preview box so it now destroys the trailer when the box is closed. If you change the trailer url and re-open the preview box it will show the updated trailer without needing to reload the page.